Moved some RecipeBuilder components into their own sub-folder to make navigation of the project a little cleaner. Updated README
